fixed threading issue
