+++ lib/jwt/decode.rb
@@ -2,6 +2,8 @@
 
 require 'json'
 
+require 'jwt/signature'
+require 'jwt/verify'
 # JWT::Decode module
 module JWT
   # Decoding logic for JWT
@@ -11,22 +13,60 @@ module JWT
       Base64.decode64(str.tr('-_', '+/'))
     end
 
-    def initialize(jwt, verify)
+    def initialize(jwt, key, verify, options, &keyfinder)
+      raise(JWT::DecodeError, 'Nil JSON web token') unless jwt
       @jwt = jwt
+      @key = key
+      @options = options
       @segments = jwt.split('.')
       @verify = verify
-      @header = ''
-      @payload = ''
       @signature = ''
+      @keyfinder = keyfinder
     end
 
     def decode_segments
       validate_segment_count
-      decode_crypto if @verify
-      return_values
+      if @verify
+        decode_crypto
+        verify_signature
+        verify_claims
+      end
+      raise(JWT::DecodeError, 'Not enough or too many segments') unless header && payload
+      [payload, header]
     end
 
     private
+
+    def verify_signature
+      @key = find_key(&@keyfinder) if @keyfinder
+
+      raise(JWT::IncorrectAlgorithm, 'An algorithm must be specified') if allowed_algorithms.empty?
+      raise(JWT::IncorrectAlgorithm, 'Expected a different algorithm') unless options_includes_algo_in_header?
+
+      Signature.verify(header['alg'], @key, signing_input, @signature)
+    end
+
+    def options_includes_algo_in_header?
+      allowed_algorithms.include? header['alg']
+    end
+
+    def allowed_algorithms
+      if @options.key?(:algorithm)
+        [@options[:algorithm]]
+      else
+        @options[:algorithms] || []
+      end
+    end
+
+    def find_key(&keyfinder)
+      key = (keyfinder.arity == 2 ? yield(header, payload) : yield(header))
+      raise JWT::DecodeError, 'No verification key available' unless key
+      key
+    end
+
+    def verify_claims
+      Verify.verify_claims(payload, @options)
+    end
 
     def validate_segment_count
       raise(JWT::DecodeError, 'Not enough or too many segments') unless
@@ -42,16 +82,12 @@ module JWT
       @signature = Decode.base64url_decode(@segments[2])
     end
 
-    def return_values
-      [header, payload, @signature, signing_input]
-    end
-
     def header
-      parse_and_decode @segments[0]
+      @header ||= parse_and_decode @segments[0]
     end
 
     def payload
-      parse_and_decode @segments[1]
+      @payload ||= parse_and_decode @segments[1]
     end
 
     def signing_input
